Fix multipolygon assembler.
Use double instead of an integer in a calculation in the multipolygon
assembler. The rounding lead to very very few cases of multipolygons
beeing build in the wrong way.

+++ include/osmium/area/detail/basic_assembler.hpp
@@ -243,17 +243,17 @@ namespace osmium {
 
                 class rings_stack_element {
 
-                    int32_t m_y;
+                    double m_y;
                     detail::ProtoRing* m_ring_ptr;
 
                 public:
 
-                    rings_stack_element(int32_t y, detail::ProtoRing* ring_ptr) :
+                    rings_stack_element(double y, detail::ProtoRing* ring_ptr) :
                         m_y(y),
                         m_ring_ptr(ring_ptr) {
                     }
 
-                    int32_t y() const noexcept {
+                    double y() const noexcept {
                         return m_y;
                     }
 
@@ -358,7 +358,7 @@ namespace osmium {
                                     std::cerr << "        Segment is below (nesting=" << nesting << ")\n";
                                 }
                                 if (segment->ring()->is_outer()) {
-                                    const int32_t y = int32_t(ay + (by - ay) * (lx - ax) / (bx - ax));
+                                    const double y = ay + (by - ay) * (lx - ax) / double(bx - ax);
                                     if (debug()) {
                                         std::cerr << "        Segment belongs to outer ring (y=" << y << " ring=" << *segment->ring() << ")\n";
                                     }
